requirements.txt
PyPDF2==3.0.1
python-dotenv==1.2.2
pandas==3.0.2
xlrd==2.0.2
markitdown==0.1.5
uvicorn==0.42.0
fastapi==0.135.2
httpx==0.28.1
langchain-openai==1.1.12
langchain-ollama==1.0.1
neo4j==6.1.0
langchain==1.2.15
langgraph==1.1.6
langchain-community==0.4.1
langchain-core==1.2.26
duckduckgo-search==8.1.1
litellm==1.83.0
python-multipart==0.0.22